==> 01_cells spatial typology.ipynb <==
import numpy as np
import pandas as pd
import matplotlib.pyplot as plt

#read 1km grid cells
gridcells1km = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_1km/geogitter/DE_Grid_ETRS89-LAEA_1km.csv', sep=';', usecols=['cellname'])
#gridcells1km

# municipality code (ags) is recorded for 250m gridcells. 
#create code for corresponding 250m gridcells
gridcells1km['corresponding250']=''

for index, row in gridcells1km.iterrows():
    gridcells1km.at[index, 'corresponding250'] = '250mN' + str(int(row['cellname'][4:8])*100+25) + 'E' + str(int(row['cellname'][9:13])*100+25)
    
#gridcells1km

#250m grids are read one after the other in distinct 100kmx100km batches.

#create empty new 1km-df to receive all the mapped 1km gridcells
gridcells1km_new = gridcells1km.iloc[0:0]

# 1-10  read 250m grid cells with ags, map the ags to a temp df, append those to the new 1km df
gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN26E43_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN26E44_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN27E41_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN27E42_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN27E43_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN27E44_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN27E45_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN28E40_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN28E41_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN28E42_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

# 11-20  read 250m grid cells with ags, map the ags to a temp df, append those to the new 1km df
gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN28E43_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN28E44_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN28E45_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN28E46_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN29E40_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN29E41_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN29E42_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN29E43_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN29E44_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN29E45_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

# 21-30  read 250m grid cells with ags, map the ags to a temp df, append those to the new 1km df
gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN30E40_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN30E41_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN30E42_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN30E43_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN30E44_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN30E45_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN30E46_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN31E40_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN31E41_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN31E42_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

# 31-40  read 250m grid cells with ags, map the ags to a temp df, append those to the new 1km df
gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN31E43_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN31E44_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN31E45_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN31E46_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN32E40_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN32E41_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN32E42_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN32E43_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN32E44_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN32E45_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

# 41-50  read 250m grid cells with ags, map the ags to a temp df, append those to the new 1km df
gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN32E46_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN33E40_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN33E41_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN33E42_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN33E43_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN33E44_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN33E45_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN33E46_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN34E40_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN34E41_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

# 51-60  read 250m grid cells with ags, map the ags to a temp df, append those to the new 1km df
gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN34E42_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN34E43_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN34E44_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN34E45_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN34E46_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN35E41_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN35E42_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN35E43_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN35E44_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells250m = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/DE_Grid_ETRS89-LAEA_250m/geogitter/100kmN35E45_DE_Grid_ETRS89-LAEA_250m.csv', low_memory=False, sep=';', usecols = [0, 11], names=['id','ags'])
gridcells1km['ags'] = gridcells1km['corresponding250'].map(gridcells250m.set_index('id')['ags'])
gridcells1km_new = pd.concat([gridcells1km_new, gridcells1km[gridcells1km['ags'].notna()]])

gridcells1km = gridcells1km_new
#gridcells1km

#replace "-" entries in ags with "0"
gridcells1km.loc[gridcells1km['ags'] == "-", 'ags'] = 0
#gridcells1km

#remove leading zeros in ags
gridcells1km['ags'] = gridcells1km['ags'].astype(float).astype(int)
gridcells1km.dtypes

#get Raumtyp (csv of ags and a two-digit code "RegioStaR4") and map it by ags
#load file
Raumtyp_dir = pd.read_csv('C:/Users/arning/sciebo/04_Research/02_Current projects/2023_04_PLUG-IN/02_Data MiD/Raumtyp.csv', low_memory=False, sep=';', skipinitialspace=True)
gridcells1km['Raumtyp'] = ''
Raumtyp_dir.dtypes
#gridcells1km.loc[gridcells1km['Raumtyp']=='']

# Create a dictionary mapping gem_20 to RegioStaR4
gem_to_RegioStaR4 = dict(zip(Raumtyp_dir['gem_20'], Raumtyp_dir['RegioStaR4']))

# Use the map function to populate the Raumtyp column in gridcells1km
gridcells1km['Raumtyp'] = gridcells1km['ags'].map(gem_to_RegioStaR4)

#Visual check if Raumtyp has been transferred to all 1km cells

#map ags by index 
#gridcells1km
#print(gridcells1km['ags'].value_counts(normalize=True))
#print(gridcells1km.notna()['ags'].value_counts(normalize=True))

#map 1km gridcells with ags using pyplot
gridcells1km_plot = gridcells1km[['Raumtyp', 'cellname']].sample(50000)
gridcells1km_plot = gridcells1km_plot[gridcells1km_plot['Raumtyp'].notna()]
gridcells1km_plot['x']=''
gridcells1km_plot['y']=''

for index, row in gridcells1km_plot.iterrows():
    gridcells1km_plot.at[index, 'x'] = int(row['cellname'][9:13])
    gridcells1km_plot.at[index, 'y'] = int(row['cellname'][4:8])

groups = gridcells1km_plot.groupby('Raumtyp')
for name, group in groups:
    plt.plot(group.x, group.y, marker='o', linestyle='', markersize=1, label=name)
plt.legend()

#gridcells1km.loc[gridcells1km['cellname']=='1kmN3301E4299']
gridcells1km[gridcells1km.isna().any(axis=1)].sample(100)

#save Raumtyp file to use in model
gridcells1km.to_csv('Raumtyp.csv')
