==> 21_nested Logit.ipynb <==
import pandas as pd
import numpy as np
import biogeme.database as db
import biogeme.biogeme as bio
from biogeme.expressions import Beta, Variable
from biogeme.models import loglogit, lognested
pd.set_option('display.max_columns', None)

# Read data file created by "Data processing"
d = pd.read_csv('D:/Processed Data/DE_ownership_input_30k.csv')
d_biogeme = d[['vpedrad','elevar', 'Raumtyp_12', 'Raumtyp_21', 'Raumtyp_22', 'alter_gr_1', 'alter_gr_2', 'alter_gr_3', 'alter_gr_5', 'alter_gr_6', 'alter_gr_7', 'alter_gr_8', 'P_BIL_1', 'P_BIL_2', 'P_BIL_3', 'P_BIL_5', 'P_BIL_6', 'HP_SEX_2', 'taet_2', 'taet_3', 'taet_4', 'taet_5', 'oek_status_1', 'oek_status_2', 'oek_status_4', 'oek_status_5', 'hhgr_gr2_2', 'hhgr_gr2_3', 'hhgr_gr2_4']].copy()
database = db.Database('ebike-ownership-MiD', d_biogeme)
#d_biogeme 

# Define variables

# vpedrad takes values from 1 to 4, one value corresponding to one of four possible bundles
choice = Variable('vpedrad')

#elevar (="gradient") is the only continous variable, all others are categorical
elevar = Variable('elevar')

Raumtyp_12 = Variable('Raumtyp_12')
Raumtyp_21 = Variable('Raumtyp_21')
Raumtyp_22 = Variable('Raumtyp_22')

alter_gr_1 = Variable('alter_gr_1')   # Used both for interaction and as contribution to utility function in its own right; for minors, age_1 effectively "replaces" the influence of occupation and level of education.
alter_gr_2 = Variable('alter_gr_2')
alter_gr_3 = Variable('alter_gr_3')
alter_gr_5 = Variable('alter_gr_5')
alter_gr_6 = Variable('alter_gr_6')
alter_gr_7 = Variable('alter_gr_7')
alter_gr_8 = Variable('alter_gr_8')

#P_BIL_1 = Variable('P_BIL_1')        # Not used despite not being reference category, since occupation and level of education are already interacted with "age =!1".
P_BIL_2 = Variable('P_BIL_2')
P_BIL_3 = Variable('P_BIL_3')
P_BIL_5 = Variable('P_BIL_5')
P_BIL_6 = Variable('P_BIL_6')

HP_SEX_2 = Variable('HP_SEX_2')

taet_2 = Variable('taet_2')          # Used despite including school students, because this category also includes university students and apprentices
taet_3 = Variable('taet_3')
taet_4 = Variable('taet_4')
taet_5 = Variable('taet_5')

oek_status_1 = Variable('oek_status_1')
oek_status_2 = Variable('oek_status_2')
oek_status_4 = Variable('oek_status_4')
oek_status_5 = Variable('oek_status_5')

hhgr_gr2_2 = Variable('hhgr_gr2_2')
hhgr_gr2_3 = Variable('hhgr_gr2_3')
hhgr_gr2_4 = Variable('hhgr_gr2_4')

# Parameters to be estimated
# Arguments:
#   1  Name for report. Typically, the same as the variable
#   2  Starting value
#   3  Lower bound
#   4  Upper bound
#   5  0: estimate the parameter, 1: keep it fixe

#ASC
ASC_cbike = Beta('ASC_cbike', 4, None, None, 0)
ASC_ebike = Beta('ASC_ebike', -25, None, None, 0)
ASC_bothbike = Beta('ASC_bothbike', -4, None, None, 0)

#betas for elevar
beta_elevar_cbike = Beta('beta_elevar_cbike', -14, None, None, 0)
beta_elevar_ebike = Beta('beta_elevar_ebike', -12, None, None, 0)
beta_elevar_bothbike = Beta('beta_elevar_bothbike', -8, None, None, 0)

#betas for Raumtyp
beta_Raumtyp_12_cbike = Beta('beta_Raumtyp_12_cbike', 0.1, None, None, 0)
beta_Raumtyp_21_cbike = Beta('beta_Raumtyp_21_cbike', 0.1, None, None, 0)
beta_Raumtyp_22_cbike = Beta('beta_Raumtyp_22_cbike', 0.4, None, None, 0)

beta_Raumtyp_12_ebike = Beta('beta_Raumtyp_12_ebike', 0.3, None, None, 0)
beta_Raumtyp_21_ebike = Beta('beta_Raumtyp_21_ebike', 2.8, None, None, 0)
beta_Raumtyp_22_ebike = Beta('beta_Raumtyp_22_ebike', 0.2, None, None, 0)

beta_Raumtyp_12_bothbike = Beta('beta_Raumtyp_12_bothbike', 0.1, None, None, 0)
beta_Raumtyp_21_bothbike = Beta('beta_Raumtyp_21_bothbike', -1, None, None, 0)
beta_Raumtyp_22_bothbike = Beta('beta_Raumtyp_22_bothbike', 0.8, None, None, 0)

#betas for alter_gr_
beta_alter_gr_1_cbike = Beta('beta_alter_gr_1_cbike', 0, None, None, 0)
beta_alter_gr_2_cbike = Beta('beta_alter_gr_2_cbike', -1, None, None, 0)
beta_alter_gr_3_cbike = Beta('beta_alter_gr_3_cbike', -0.5, None, None, 0)
beta_alter_gr_5_cbike = Beta('beta_alter_gr_5_cbike', 0, None, None, 0)
beta_alter_gr_6_cbike = Beta('beta_alter_gr_6_cbike', -0.4, None, None, 0)
beta_alter_gr_7_cbike = Beta('beta_alter_gr_7_cbike', -1.3, None, None, 0)
beta_alter_gr_8_cbike = Beta('beta_alter_gr_8_cbike', -3, None, None, 0)

beta_alter_gr_1_ebike = Beta('beta_alter_gr_1_ebike', -5.2, None, None, 0)
beta_alter_gr_2_ebike = Beta('beta_alter_gr_2_ebike', -8, None, None, 0)
beta_alter_gr_3_ebike = Beta('beta_alter_gr_3_ebike', 3.3, None, None, 0)
beta_alter_gr_5_ebike = Beta('beta_alter_gr_5_ebike', 3.5, None, None, 0)
beta_alter_gr_6_ebike = Beta('beta_alter_gr_6_ebike', 5.5, None, None, 0)
beta_alter_gr_7_ebike = Beta('beta_alter_gr_7_ebike', 2.3, None, None, 0)
beta_alter_gr_8_ebike = Beta('beta_alter_gr_8_ebike', 2.2, None, None, 0)

beta_alter_gr_1_bothbike = Beta('beta_alter_gr_1_bothbike', -7.9, None, None, 0)
beta_alter_gr_2_bothbike = Beta('beta_alter_gr_2_bothbike', 0, None, None, 0)
beta_alter_gr_3_bothbike = Beta('beta_alter_gr_3_bothbike', -3, None, None, 0)
beta_alter_gr_5_bothbike = Beta('beta_alter_gr_5_bothbike', -0.5, None, None, 0)
beta_alter_gr_6_bothbike = Beta('beta_alter_gr_6_bothbike', -1.9, None, None, 0)
beta_alter_gr_7_bothbike = Beta('beta_alter_gr_7_bothbike', -1, None, None, 0)
beta_alter_gr_8_bothbike = Beta('beta_alter_gr_8_bothbike', -3.7, None, None, 0)

#betas for P_BIL_
#beta_P_BIL_1_cbike = Beta('beta_P_BIL_1_cbike', 0, None, None, 0)
beta_P_BIL_2_cbike = Beta('beta_P_BIL_2_cbike', -0.4, None, None, 0)
beta_P_BIL_3_cbike = Beta('beta_P_BIL_3_cbike', -0.2, None, None, 0)
beta_P_BIL_5_cbike = Beta('beta_P_BIL_5_cbike', 0.5, None, None, 0)
beta_P_BIL_6_cbike = Beta('beta_P_BIL_6_cbike', -0.5, None, None, 0)

#beta_P_BIL_1_ebike = Beta('beta_P_BIL_1_ebike', 0, None, None, 0)
beta_P_BIL_2_ebike = Beta('beta_P_BIL_2_ebike', 5, None, None, 0)
beta_P_BIL_3_ebike = Beta('beta_P_BIL_3_ebike', 2.8, None, None, 0)
beta_P_BIL_5_ebike = Beta('beta_P_BIL_5_ebike', -0.5, None, None, 0)
beta_P_BIL_6_ebike = Beta('beta_P_BIL_6_ebike', -4.4, None, None, 0)

#beta_P_BIL_1_bothbike = Beta('beta_P_BIL_1_bothbike', 0, None, None, 0)
beta_P_BIL_2_bothbike = Beta('beta_P_BIL_2_bothbike', -3.8, None, None, 0)
beta_P_BIL_3_bothbike = Beta('beta_P_BIL_3_bothbike', -2, None, None, 0)
beta_P_BIL_5_bothbike = Beta('beta_P_BIL_5_bothbike', 0.7, None, None, 0)
beta_P_BIL_6_bothbike = Beta('beta_P_BIL_6_bothbike', 0.9, None, None, 0)

#betas for HP_SEX_
beta_HP_SEX_2_cbike = Beta('beta_HP_SEX_2_cbike', -0.6, None, None, 0)
beta_HP_SEX_2_ebike = Beta('beta_HP_SEX_2_ebike', -0.1, None, None, 0)
beta_HP_SEX_2_bothbike = Beta('beta_HP_SEX_2_bothbike', -0.9, None, None, 0)

#betas for taet_
beta_taet_2_cbike = Beta('beta_taet_2_cbike', 0.5, None, None, 0)
beta_taet_3_cbike = Beta('beta_taet_3_cbike', -1, None, None, 0)
beta_taet_4_cbike = Beta('beta_taet_4_cbike', -0.8, None, None, 0)
beta_taet_5_cbike = Beta('beta_taet_5_cbike', -1, None, None, 0)

beta_taet_2_ebike = Beta('beta_taet_2_ebike', -28.7, None, None, 0)
beta_taet_3_ebike = Beta('beta_taet_3_ebike', 2.6, None, None, 0)
beta_taet_4_ebike = Beta('beta_taet_4_ebike', 1.2, None, None, 0)
beta_taet_5_ebike = Beta('beta_taet_5_ebike', 5.7, None, None, 0)

beta_taet_2_bothbike = Beta('beta_taet_2_bothbike', 2.3, None, None, 0)
beta_taet_3_bothbike = Beta('beta_taet_3_bothbike', -3, None, None, 0)
beta_taet_4_bothbike = Beta('beta_taet_4_bothbike', -1.2, None, None, 0)
beta_taet_5_bothbike = Beta('beta_taet_5_bothbike', -5.8, None, None, 0)

#betas for oek_status_
beta_oek_status_1_cbike = Beta('beta_oek_status_1_cbike', -0.4, None, None, 0)
beta_oek_status_2_cbike = Beta('beta_oek_status_2_cbike', -0.2, None, None, 0)
beta_oek_status_4_cbike = Beta('beta_oek_status_4_cbike', 1, None, None, 0)
beta_oek_status_5_cbike = Beta('beta_oek_status_5_cbike', 0.8, None, None, 0)

beta_oek_status_1_ebike = Beta('beta_oek_status_1_ebike', 0.9, None, None, 0)
beta_oek_status_2_ebike = Beta('beta_oek_status_2_ebike', -1.7, None, None, 0)
beta_oek_status_4_ebike = Beta('beta_oek_status_4_ebike', 2, None, None, 0)
beta_oek_status_5_ebike = Beta('beta_oek_status_5_ebike', 1.7, None, None, 0)

beta_oek_status_1_bothbike = Beta('beta_oek_status_1_bothbike', -3, None, None, 0)
beta_oek_status_2_bothbike = Beta('beta_oek_status_2_bothbike', 0, None, None, 0)
beta_oek_status_4_bothbike = Beta('beta_oek_status_4_bothbike', 0.3, None, None, 0)
beta_oek_status_5_bothbike = Beta('beta_oek_status_5_bothbike', -0.1, None, None, 0)

#betas for hhgr_gr2_
beta_hhgr_gr2_2_cbike = Beta('beta_hhgr_gr2_2_cbike', 0.8, None, None, 0)
beta_hhgr_gr2_3_cbike = Beta('beta_hhgr_gr2_3_cbike', 0.9, None, None, 0)
beta_hhgr_gr2_4_cbike = Beta('beta_hhgr_gr2_4_cbike', 1.9, None, None, 0)

beta_hhgr_gr2_2_ebike = Beta('beta_hhgr_gr2_2_ebike', 2.6, None, None, 0)
beta_hhgr_gr2_3_ebike = Beta('beta_hhgr_gr2_3_ebike', -0.4, None, None, 0)
beta_hhgr_gr2_4_ebike = Beta('beta_hhgr_gr2_4_ebike', -2.3, None, None, 0)

beta_hhgr_gr2_2_bothbike = Beta('beta_hhgr_gr2_2_bothbike', 0, None, None, 0)
beta_hhgr_gr2_3_bothbike = Beta('beta_hhgr_gr2_3_bothbike', 1.4, None, None, 0)
beta_hhgr_gr2_4_bothbike = Beta('beta_hhgr_gr2_4_bothbike', 3.2, None, None, 0)

# Utilities
nobike = (
    0
)

cbike = (
    ASC_cbike +
    beta_elevar_cbike * elevar +
    beta_Raumtyp_12_cbike * Raumtyp_12 +
    beta_Raumtyp_21_cbike * Raumtyp_21 +
    beta_Raumtyp_22_cbike * Raumtyp_22 +
    beta_alter_gr_1_cbike * alter_gr_1 + 
    beta_alter_gr_2_cbike * alter_gr_2 +
    beta_alter_gr_3_cbike * alter_gr_3 +
    beta_alter_gr_5_cbike * alter_gr_5 +
    beta_alter_gr_6_cbike * alter_gr_6 +
    beta_alter_gr_7_cbike * alter_gr_7 +
    beta_alter_gr_8_cbike * alter_gr_8 +
    beta_P_BIL_2_cbike * P_BIL_2 * (1-alter_gr_1) +
    beta_P_BIL_3_cbike * P_BIL_3 * (1-alter_gr_1) +
    beta_P_BIL_5_cbike * P_BIL_5 * (1-alter_gr_1) +
    beta_P_BIL_6_cbike * P_BIL_6 * (1-alter_gr_1) +
    beta_HP_SEX_2_cbike * HP_SEX_2 +
    beta_taet_2_cbike * taet_2 * (1-alter_gr_1) +
    beta_taet_3_cbike * taet_3 * (1-alter_gr_1) +
    beta_taet_4_cbike * taet_4 * (1-alter_gr_1) +
    beta_taet_5_cbike * taet_5 * (1-alter_gr_1) +
    beta_oek_status_1_cbike * oek_status_1 +
    beta_oek_status_2_cbike * oek_status_2 +
    beta_oek_status_4_cbike * oek_status_4 +
    beta_oek_status_5_cbike * oek_status_5 +
    beta_hhgr_gr2_2_cbike * hhgr_gr2_2 +
    beta_hhgr_gr2_3_cbike * hhgr_gr2_3 +
    beta_hhgr_gr2_4_cbike * hhgr_gr2_4
)

ebike = (
    ASC_ebike +
    beta_elevar_ebike * elevar +
    beta_Raumtyp_12_ebike * Raumtyp_12 +
    beta_Raumtyp_21_ebike * Raumtyp_21 +
    beta_Raumtyp_22_ebike * Raumtyp_22 +
    beta_alter_gr_1_ebike * alter_gr_1 +
    beta_alter_gr_2_ebike * alter_gr_2 +
    beta_alter_gr_3_ebike * alter_gr_3 +
    beta_alter_gr_5_ebike * alter_gr_5 +
    beta_alter_gr_6_ebike * alter_gr_6 +
    beta_alter_gr_7_ebike * alter_gr_7 +
    beta_alter_gr_8_ebike * alter_gr_8 +
    beta_P_BIL_2_ebike * P_BIL_2 * (1-alter_gr_1) +
    beta_P_BIL_3_ebike * P_BIL_3 * (1-alter_gr_1) +
    beta_P_BIL_5_ebike * P_BIL_5 * (1-alter_gr_1) +
    beta_P_BIL_6_ebike * P_BIL_6 * (1-alter_gr_1) +
    beta_HP_SEX_2_ebike * HP_SEX_2 +
    beta_taet_2_ebike * taet_2 * (1-alter_gr_1) +
    beta_taet_3_ebike * taet_3 * (1-alter_gr_1) +
    beta_taet_4_ebike * taet_4 * (1-alter_gr_1) +
    beta_taet_5_ebike * taet_5 * (1-alter_gr_1) +
    beta_oek_status_1_ebike * oek_status_1 +
    beta_oek_status_2_ebike * oek_status_2 +
    beta_oek_status_4_ebike * oek_status_4 +
    beta_oek_status_5_ebike * oek_status_5 +
    beta_hhgr_gr2_2_ebike * hhgr_gr2_2 +
    beta_hhgr_gr2_3_ebike * hhgr_gr2_3 +
    beta_hhgr_gr2_4_ebike * hhgr_gr2_4
)

bothbike = (
    ASC_bothbike +
    beta_elevar_bothbike * elevar +
    beta_Raumtyp_12_bothbike * Raumtyp_12 +
    beta_Raumtyp_21_bothbike * Raumtyp_21 +
    beta_Raumtyp_22_bothbike * Raumtyp_22 +
    beta_alter_gr_1_bothbike * alter_gr_1 + 
    beta_alter_gr_2_bothbike * alter_gr_2 +
    beta_alter_gr_3_bothbike * alter_gr_3 +
    beta_alter_gr_5_bothbike * alter_gr_5 +
    beta_alter_gr_6_bothbike * alter_gr_6 +
    beta_alter_gr_7_bothbike * alter_gr_7 +
    beta_alter_gr_8_bothbike * alter_gr_8 +
    beta_P_BIL_2_bothbike * P_BIL_2 * (1-alter_gr_1) +
    beta_P_BIL_3_bothbike * P_BIL_3 * (1-alter_gr_1) +
    beta_P_BIL_5_bothbike * P_BIL_5 * (1-alter_gr_1) +
    beta_P_BIL_6_bothbike * P_BIL_6 * (1-alter_gr_1) +
    beta_HP_SEX_2_bothbike * HP_SEX_2 +
    beta_taet_2_bothbike * taet_2 * (1-alter_gr_1) +
    beta_taet_3_bothbike * taet_3 * (1-alter_gr_1) +
    beta_taet_4_bothbike * taet_4 * (1-alter_gr_1) +
    beta_taet_5_bothbike * taet_5 * (1-alter_gr_1) +
    beta_oek_status_1_bothbike * oek_status_1 +
    beta_oek_status_2_bothbike * oek_status_2 +
    beta_oek_status_4_bothbike * oek_status_4 +
    beta_oek_status_5_bothbike * oek_status_5 +
    beta_hhgr_gr2_2_bothbike * hhgr_gr2_2 +
    beta_hhgr_gr2_3_bothbike * hhgr_gr2_3 +
    beta_hhgr_gr2_4_bothbike * hhgr_gr2_4
)

V = {1: cbike, 2: ebike, 3: bothbike, 4: nobike}
av = {1: 1, 2: 1, 3: 1, 4: 1}

# Nest definition
N_ebike_yes = Beta('N_ebike_yes', 1, 1, 100, 0)
N_ebikeno_cbikeyes = 1
N_ebikeno_cbikeno = 1

nest_ebike_yes = (N_ebike_yes, [2, 3])
nest_ebikeno_cbikeyes = (N_ebikeno_cbikeyes, [1])
nest_ebikeno_cbikeno = (N_ebikeno_cbikeno, [4])
nests = nest_ebike_yes, nest_ebikeno_cbikeno, nest_ebikeno_cbikeyes

# Nested Logit model
logprob = lognested(V, av, nests, choice)
biogeme = bio.BIOGEME(database, logprob)
#biogeme.algorithm = 'simple_bounds'  # LS-BFGS-B algorithm
#biogeme.algorithm = 'newton'       # Newton method
#biogeme.algorithm = 'simplex'      # Nelder-Mead Simplex method
biogeme.modelName = 'ebike-ownership-MiD_NL_revised'
biogeme.calculateNullLoglikelihood(av)
results = biogeme.estimate()

# Multinomial Logit model
# logprob = loglogit(V, av, choice)
# biogeme = bio.BIOGEME(database, logprob)
# biogeme.modelName = 'ebike-ownership-MiD'
# biogeme.calculateNullLoglikelihood(av)
# results = biogeme.estimate()
